Clarinet check working Initial test file and adoption of vitest and new clarinet sdk
