--- tests/contract.test.ts
@@ -1,20 +1,0 @@
-import { describe, expect, it } from "vitest";
-
-const accounts = simnet.getAccounts();
-const address1 = accounts.get("wallet_1")!;
-
-/*
-  The test below is an example. Learn more in the clarinet-sdk readme:
-  https://github.com/hirosystems/clarinet/blob/develop/components/clarinet-sdk/README.md
-*/
-
-describe("example tests", () => {
-  it("ensures simnet is well initalise", () => {
-    expect(simnet.blockHeight).toBe(0);
-  });
-
-  // it("shows an example", () => {
-  //   const { result } = simnet.callReadOnlyFn("counter", "get-counter", [], address1);
-  //   expect(result).toBeUint(0);
-  // });
-});
