+++ tsconfig.json
@@ -21,5 +21,5 @@
   "include": [
     "node_modules/@hirosystems/clarinet-sdk/vitest-helpers/src",
     "tests"
-  ]
+, "utils/old"  ]
 }
