--- utils/mine_btc.sh
@@ -1,19 +1,0 @@
-#!/bin/bash
-
-# Mines some BTC to the default BTC address
-
-num_blocks=$1
-dir="$(dirname "$0")"
-
-if [[ -z "$num_blocks" ]]; then
-    echo "Usage: ./script.sh [num_blocks]"
-    exit 1
-fi
-
-btc_address="bcrt1q3tj2fr9scwmcw3rq5m6jslva65f2rqjxfrjz47"
-#$(source $dir/get_credentials.sh | jq -r '.credentials["0"].bitcoin.p2wpkh.address')
-$dir/../bitcoin/bin/bitcoin-cli generatetoaddress 1 $btc_address
-echo "Mined 1 BTC to $btc_address"
-btc_address=$(source $dir/get_credentials.sh | jq -r '.credentials["1"].bitcoin.p2wpkh.address')
-$dir/../bitcoin/bin/bitcoin-cli generatetoaddress $num_blocks $btc_address
-echo "Mined $num_blocks BTC to $btc_address"
